feat: active style and some header of layout
